食谱、口味、用料、步骤、功效表添加保存操作

+++ 代码/paichi/src/main/java/com/paichi/modules/craft/entity/Craft.java
@@ -28,7 +28,7 @@ public class Craft implements Serializable {
      * 食谱工艺表主键
      */
     @TableId("CRAFT_ID")
-    private Double craftId;
+    private Integer craftId;
 
     /**
      * 食谱工艺名称
